move code inside windows
