+++ crates/gpui/src/platform/windows.rs
@@ -1,4 +1,5 @@
 mod clipboard;
+mod destination_list;
 mod direct_write;
 mod dispatcher;
 mod display;
@@ -10,6 +11,7 @@ mod window;
 mod wrapper;
 
 pub(crate) use clipboard::*;
+pub(crate) use destination_list::*;
 pub(crate) use direct_write::*;
 pub(crate) use dispatcher::*;
 pub(crate) use display::*;
